move to CvGameCoreUtils
move to CvGameCoreUtils
