+++ CvGameCoreDLL_Expansion2/VoxPopuli.vcxproj.filters
@@ -310,7 +310,6 @@
     <ClCompile Include="Lua\CvLuaUnit.cpp">
       <Filter>interface</Filter>
     </ClCompile>
-    <ClCompile Include="CvAssertDlg.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\CvGameCoreDLLUtil\include\CvEnums.h" />
