+++ build_vp_clang_sdk.py
@@ -120,7 +120,6 @@ CPP = [
     'CvGameCoreDLL_Expansion2\\CvAIOperation.cpp',
     'CvGameCoreDLL_Expansion2\\CvArea.cpp',
     'CvGameCoreDLL_Expansion2\\CvArmyAI.cpp',
-    'CvGameCoreDLL_Expansion2\\CvAssertDlg.cpp',
     'CvGameCoreDLL_Expansion2\\CvAStar.cpp',
     'CvGameCoreDLL_Expansion2\\CvAStarNode.cpp',
     'CvGameCoreDLL_Expansion2\\CvBarbarians.cpp',
@@ -343,7 +342,7 @@ def build_cl_config_args(config: Config) -> list[str]:
     return args
 
 def build_link_config_args(config: Config) -> list[str]:
-    args = ['/MACHINE:x86', '/DLL', '/DEBUG', '/LTCG', '/DYNAMICBASE', '/NXCOMPAT', '/SUBSYSTEM:WINDOWS', '/MANIFEST:EMBED', f'/DEF:"{os.path.join(PROJECT_DIR, DEF_FILE)}"']
+    args = ['/MACHINE:x86', '/DLL', '/DEBUG', '/LTCG', '/DYNAMICBASE', '/NXCOMPAT', '/SUBSYSTEM:WINDOWS', '/MANIFEST:EMBED', '/FORCE:MULTIPLE', f'/DEF:"{os.path.join(PROJECT_DIR, DEF_FILE)}"']
     if config == Config.Release:
         args += ['/OPT:REF', '/OPT:ICF']
     return args
